# Add serialization copy benchmark, optimize copying and reading with IL generators