--- a/Robust.Benchmarks/Serialization/Copy/SerializationCopyBenchmark.cs
+++ b/Robust.Benchmarks/Serialization/Copy/SerializationCopyBenchmark.cs
@@ -87,6 +87,7 @@ namespace Robust.Benchmarks.Serialization.Copy
         [Benchmark]
         public SeedDataDefinition BaselineCreateCopySeedDataDefinition()
         {
+            // ReSharper disable once UseObjectOrCollectionInitializer
             var copy = new SeedDataDefinition();
 
             copy.ID = Seed.ID;