--- a/Robust.Benchmarks/Serialization/Definitions/DataDefinitionWithString.cs
+++ b/Robust.Benchmarks/Serialization/Definitions/DataDefinitionWithString.cs
@@ -1,6 +1,6 @@
 ﻿using Robust.Shared.Serialization.Manager.Attributes;
 
-namespace Robust.Benchmarks.Serialization
+namespace Robust.Benchmarks.Serialization.Definitions
 {
     [DataDefinition]
     public class DataDefinitionWithString