--- a/Robust.Shared/Serialization/Manager/SerializationDataDefinition.cs
+++ b/Robust.Shared/Serialization/Manager/SerializationDataDefinition.cs
@@ -47,6 +47,8 @@ namespace Robust.Shared.Serialization.Manager
             object value,
             DeserializedFieldEntry[] mappings);
 
+        private delegate TValue AccessField<TTarget, TValue>(ref TTarget target);
+
         private delegate void AssignField<TTarget, TValue>(ref TTarget target, TValue? value);
 
         public readonly Type Type;
@@ -59,6 +61,7 @@ namespace Robust.Shared.Serialization.Manager
         private readonly SerializeDelegateSignature _serializeDelegate;
         private readonly CopyDelegateSignature _copyDelegate;
 
+        private readonly AccessField<object, object?>[] _fieldAccessors;
         private readonly AssignField<object, object?>[] _fieldAssigners;
 
         public SerializationDataDefinition(Type type)
@@ -142,6 +145,60 @@ namespace Robust.Shared.Serialization.Manager
             _populateDelegate = EmitPopulateDelegate();
             _serializeDelegate = EmitSerializeDelegate();
             _copyDelegate = EmitCopyDelegate();
+
+            _fieldAccessors = new AccessField<object, object?>[BaseFieldDefinitions.Length];
+
+            for (var i = 0; i < BaseFieldDefinitions.Length; i++)
+            {
+                var fieldDefinition = BaseFieldDefinitions[i];
+                var dm = new DynamicMethod(
+                    "AccessField",
+                    typeof(object),
+                    new[] {typeof(object).MakeByRefType()},
+                    true);
+
+                dm.DefineParameter(1, ParameterAttributes.Out, "target");
+
+                var generator = dm.GetRobustGen();
+
+                if (Type.IsValueType)
+                {
+                    generator.DeclareLocal(Type);
+                    generator.Emit(OpCodes.Ldarg_0);
+                    generator.Emit(OpCodes.Ldind_Ref);
+                    generator.Emit(OpCodes.Unbox_Any, Type);
+                    generator.Emit(OpCodes.Stloc_0);
+                    generator.Emit(OpCodes.Ldloca_S, 0);
+                }
+                else
+                {
+                    generator.Emit(OpCodes.Ldarg_0);
+                    generator.Emit(OpCodes.Ldind_Ref);
+                    generator.Emit(OpCodes.Castclass, Type);
+                }
+
+                switch (fieldDefinition.BackingField)
+                {
+                    case SpecificFieldInfo field:
+                        generator.Emit(OpCodes.Ldfld, field.FieldInfo);
+                        break;
+                    case SpecificPropertyInfo property:
+                        var getter = property.PropertyInfo.GetGetMethod(true) ?? throw new NullReferenceException();
+                        var opCode = Type.IsValueType ? OpCodes.Call : OpCodes.Callvirt;
+                        generator.Emit(opCode, getter);
+                        break;
+                }
+
+                var returnType = fieldDefinition.BackingField.FieldType;
+                if (returnType.IsValueType)
+                {
+                    generator.Emit(OpCodes.Box, returnType);
+                }
+
+                generator.Emit(OpCodes.Ret);
+
+                _fieldAccessors[i] = dm.CreateDelegate<AccessField<object, object?>>();
+            }
 
             _fieldAssigners = new AssignField<object, object?>[BaseFieldDefinitions.Length];
 
@@ -355,7 +412,6 @@ namespace Robust.Shared.Serialization.Manager
                     var res = deserializedFields[i];
                     if (!res.Mapped) continue;
 
-                    var fieldDefinition = BaseFieldDefinitions[i];
                     var defValue = defaultValues[i];
 
                     if (Equals(res.Result?.RawValue, defValue))
@@ -432,7 +488,7 @@ namespace Robust.Shared.Serialization.Manager
         // todo paul add skiphooks
         private CopyDelegateSignature EmitCopyDelegate()
         {
-            object PopulateDelegate(
+            object CopyDelegate(
                 object source,
                 object target,
                 ISerializationManager manager,
@@ -441,8 +497,9 @@ namespace Robust.Shared.Serialization.Manager
                 for (var i = 0; i < BaseFieldDefinitions.Length; i++)
                 {
                     var field = BaseFieldDefinitions[i];
-                    var sourceValue = field.GetValue(source);
-                    var targetValue = field.GetValue(target);
+                    var accessor = _fieldAccessors[i];
+                    var sourceValue = accessor(ref source);
+                    var targetValue = accessor(ref target);
 
                     object? copy;
                     if (sourceValue != null &&
@@ -466,7 +523,7 @@ namespace Robust.Shared.Serialization.Manager
                 return target;
             }
 
-            return PopulateDelegate;
+            return CopyDelegate;
         }
 
         public class FieldDefinition