--- a/Robust.Shared/Serialization/Manager/SerializationManager.TypeReaders.cs
+++ b/Robust.Shared/Serialization/Manager/SerializationManager.TypeReaders.cs
@@ -21,7 +21,7 @@ namespace Robust.Shared.Serialization.Manager
 
         private readonly ConcurrentDictionary<Type, ConcurrentDictionary<Type, ReadDelegate>> _readDelegates = new();
 
-        private ReadDelegate GetOrCreateDelegate(Type type, Type nodeType)
+        private ReadDelegate GetOrCreateReadDelegate(Type type, Type nodeType)
         {
             return _readDelegates
                 .GetOrAdd(type, _ => new ConcurrentDictionary<Type, ReadDelegate>())
@@ -105,7 +105,7 @@ namespace Robust.Shared.Serialization.Manager
             bool skipHook,
             ISerializationContext? context = null)
         {
-            return GetOrCreateDelegate(type, node.GetType())(type, node, dependencies, out obj, skipHook, context);
+            return GetOrCreateReadDelegate(type, node.GetType())(type, node, dependencies, out obj, skipHook, context);
         }
 
         private bool TryGetGenericReader<T, TNode>(