… Robust.Shared/Serialization/Manager/SerializationManager.TypeCopiers.cs
… Robust.Shared/Serialization/Manager/SerializationManager.cs
… Robust.Shared/Utility/ILGeneratorExt.cs
… Robust.Shared/Utility/TypeHelpers.cs