Copying Before
|                             Method |         Mean |      Error |     StdDev |       Median |
|----------------------------------- |-------------:|-----------:|-----------:|-------------:|
|                   CreateCopyString |     43.54 ns |   0.918 ns |   0.859 ns |     43.74 ns |
|                  CreateCopyInteger |     19.33 ns |   0.388 ns |   0.860 ns |     19.24 ns |
| CreateCopyDataDefinitionWithString |  3,971.42 ns |  78.974 ns | 213.512 ns |  3,902.54 ns |
|       CreateCopySeedDataDefinition | 49,916.69 ns | 264.992 ns | 221.280 ns | 49,950.53 ns |

Copying After
|                               Method |        Mean |      Error |     StdDev |
|------------------------------------- |------------:|-----------:|-----------:|
|                     CreateCopyString |    43.69 ns |   0.659 ns |   0.550 ns |
|                    CreateCopyInteger |    19.06 ns |   0.393 ns |   0.551 ns |
|   CreateCopyDataDefinitionWithString |   411.91 ns |   7.495 ns |   8.923 ns |
|         CreateCopySeedDataDefinition | 6,636.86 ns | 132.316 ns | 281.976 ns |


[Reading Before](https://github.com/space-wizards/RobustToolbox/pull/1679)
|                 Method |        Mean |     Error |    StdDev |
|----------------------- |------------:|----------:|----------:|
|             ReadString |    227.1 ns |   4.47 ns |   5.65 ns |
|            ReadInteger |    245.4 ns |   4.82 ns |   6.26 ns |
|  ReadDataDefWithString |    804.7 ns |  15.27 ns |  16.34 ns |
| ReadSeedDataDefinition | 15,846.8 ns | 312.89 ns | 773.39 ns |

Reading After
|                       Method |        Mean |     Error |    StdDev |
|----------------------------- |------------:|----------:|----------:|
|                   ReadString |    202.8 ns |   2.43 ns |   2.27 ns |
|                  ReadInteger |    266.1 ns |   5.19 ns |   7.11 ns |
| ReadDataDefinitionWithString |    596.5 ns |   4.56 ns |   3.56 ns |
|       ReadSeedDataDefinition | 11,689.2 ns | 211.90 ns | 187.85 ns |



Populate and copy delegates have been changed to emit IL code, copyByRef is now a hashset and the copy method lookup has been changed to use compiled expressions.